requirements.txt
flask
flask_sqlalchemy
datetime
codecs
ecdsa
pycryptodome